Costume, PWT, and Ubers, oh my!
Back at it again with some new sets!
This repo adds 9 Pokemon to the Costume set, and 7 sets to the PWT set (including the first Gen 8 series in the form of Macro Cosmos's), as well as 1 update to the Team Skull series that it desperately needed (seriously, Koffing? Since when does Team Skull use Koffing????). The set also adds some minor updates to 3 uber sets in the form of a few new moves gained during the Crown Tundra update.