Small Adjustments
I know I said I was done, but I saw a comment saying that Triffid was not changed from the HW version and it just... ballooned from there.

* Modified Triffid's stats to match its Apokelypse counterpart. I guess it's a nerf.
* Added a few extra HW/XM costumes (before you ask, yes they're references)
* Added two PWT movesets so I can contribute at least something that can be easily addable. Note that for Anabel, each combination refers to a different variation of that Poke's moveset

I would've added her Sun/Moon set, but that includes some moves that aren't in Gen IV and I'm still not sure how accurate we should be to the original (some movesets here include Team Rocket's anime versions and Morgan having a Stantler named Sawsbuck). If we really need to be as close as possible, expect more Colosseum/XD/PBR sets with further updates.