PWT Additions
I know I said this was the last time, but I noticed an error and it ballooned from there. Main changes:

* Sinnoh Elite 4 + Cynthia - Noticed that they get their movesets from Platinum, so I also added the sets from Diamond and Pearl. Used combinations to ensure that only one or the other comes up at any one time.
* Hoenn Elite Four + Steven - The same thing, adding in their movesets from Ruby and Sapphire as well as Omega Ruby and Alpha Sapphire. Again, used separations and combinations to divide them up. Also, there was a folder for Sidney and another for Sydney for some reason.
* Additions: Added a few more characters from the Let's Go games, as well as some Gen 8 representaion. Finally, tried out a little something new where each Pokemon in a set corresponds to a different trainer - one for Raid Partners, the other for Contest NPCs.

And yes, some movesets had to be tweaked a bit because they contain moves or forms not in Gen IV. I know that authenticity is usually the rule here, but I also saw that Morgan has a Stantler called "Sawsbuck" and Nanu has a Persian with BlackGlasses and Thick Fat, so clearly that rule is flexible. We can always edit them if PBR modding manages to get those moves in.

Besides that, I'm hard-pressed to find any more PWT sets to add in. Was considering adding Spark, Blanche, and Candela, but those would be just their training teams with no abilities and just whatevr attacks they use in Go. Besides that, Colosseum and XD trainers? Dunno.