Some PWT additions
Since we're waiting on Blazing to finish, why not?

* Most of the PWT sets from brawlitup99's pull request (since those were already vetted to have Gen 1-4 material)
* Some of the sets from my own pull request before (I removed all Pokemon which had Gen 5 and above abilities and moves, even if before I just tried to use an equivalent like the Tournament sets before)

We can add the other sets as necessary through approval, but these are all the official sets I could find that 1)have three or more Pokemon and 2)don't contain anything from Gen 5 and up (only caveat are the Mega Stones/Z Crystals in the Battle Tree set, but neither are functional in PBR anyway).

Also I updated the Alola E4 sets because according to Bulbapedia their Pokemon are kept in Ultra Balls and we gotta be authentic