Reverted the HW and XM set changes
As of right now, the only changes included here are the PWT additions and the runmons for Mystery Dungeon DX.

To the moveset team, I'm sorry for posting too much. I wanted to help contribute to TPP, and I guess I went a bit overboard with some of the additions, given we haven't had an update in ages. Feel free to do what you want with these changes/additions - you can also delete the other pull I had up, since that was just a less refined version of this one.