Test Your Luck runmons
Added both of them with the info M4 provided. Put them in their own file for now, although if it makes better sense I can always move them to the Sword runmons. I added Grimmsnarl just because the data was already there, but commented him out in the offchance PBR ever mods him in.

Only other change was made to costumes - after some thinking, I moved the Falinks set to Xmas because it's a single-stage Pokemon and doesn't really fit the format like the Galar starters or Kubfu. Added a new Costume mon as a replacement.