DBZ Team Training Mons
Finally found the page on the site that includes the API from the game, so I was able to add their stats and everything. I put comments next to the game-exclusive stuff for reference.

In case you're wondering:

Buuhan: No Fighting/Normal Pokemon, so I went with Lucario because that's the closest DBZ-like
Gold Frieza: No Dark/Metal mons in PBR, so I went with Mewtwo. Same reasoning as above.
Super Piccolo: Grass/Fighting, natch.
C-13: Pure Steel. Registeel is too... non-Anrdoid-y, so I went for one of the more humanoid Steel types for this gen
Imperfect Cell: Bug/Fighting, also natch
Garlic 2: Pure Dark - I could go with Mightyena, but Cacturne is more humanoid and resembles him more.
Bioman: Steel/Normal, nobody ever used it so he's the joke pick

Also I did a bit of cleaning for the PWT matches, added another set with three mons, and fixed a typo in a filename. Will save the Sanqui and Charity Mons for a later date