PWT Update
Someone on the Discord noticed that Blaine's PWT set was off (he had Slugma in his sets instead of Magcargo) and was going to joke about it until I realized that his HGSS team actually had held items that weren't added in already (according to Bulbapedia). In fact, most major FRLG and HGSS battles were missing some held item information (usually a Sitrus Berry for their ringer). I can see if it was omitted due to balancing issues, but if we're going for accuracy here in PBR we might as well go the full mile.

Pretty sure I got all of them, but let me know if I missed one.