Update Starter-Battle.yaml
So the existing sets (aka the Bulbasaur and Charmander lines) remain untouched. What I did is I simply added sets to the gimmick. Squirtle was missing, so I added it, then Johto, Hoenn and Sinnoh. I also made sure that everything was in the correct chronological order according to the national pokedex.